// frontend/src/app/models/auth.model.ts
export interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  userId: string;
  email: string;
  firstName: string;
  lastName: string;
}

export interface User {
  userId: string;
  email: string;
  firstName: string;
  lastName: string;
}